Broke stm32g4x_ha_tim into  a c and h file, otherwise struct instances  would be multiply defined
